Set up a CMake project
A complete CMake project, which works well with exercism tests.

Some of the 'c' files had to be given main() function, for it to be
possible to compile them. Normally it's not necessary for them to have a
main() function (only used as libraries for exercism tests) but in some
of them, I wrote my cusom tests.
Because of that and the fact that I wanted to keep CMakeLists files the
same for each exercise, the main() method had to be added.